fix: address review comments
- Fix evicted counter double-counting: rejected Admit blocks are NOT
  evictions (they were never in the cache), use saturating_sub(1)
- Validate --parallel-scan-cache-admission CLI value instead of
  silently mapping unknown values to Admit
- Fix 32-bit overflow risk in shard planner: cast to u64 before
  cumulative weight multiplication
- Keep Unix-only APIs unconditional (project is Linux-only)

+++ kv-engine/src/lsm_storage.rs
@@ -2378,7 +2378,8 @@ impl LsmStorageInner {
             for (key, weight) in &weighted_boundaries {
                 cumulative += *weight;
                 while next_target_idx <= max_splits
-                    && cumulative * (max_splits + 1) >= total_weight * next_target_idx
+                    && cumulative as u64 * (max_splits + 1) as u64
+                        >= total_weight as u64 * next_target_idx as u64
                 {
                     if selected.last() != Some(key) {
                         selected.push(key.clone());
